requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Signal processing
pywavelets>=1.1.1
neurokit2>=0.1.7
biosppy>=0.7.3

# Machine learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
imbalanced-learn>=0.9.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0
tensorflow>=2.8.0

# Interpretability
shap>=0.40.0
lime>=0.2.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0
joblib>=1.1.0
openpyxl>=3.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
